+++ raiden-dapp/src/components/navigation/Channels.vue
@@ -166,7 +166,8 @@ export default class Channels extends Mixins(NavigationMixin) {
 .channels {
   width: 100%;
   height: 100%;
-  overflow: hidden;
+  overflow-y: scroll;
+  @extend .themed-scrollbar;
 
   &:first-child {
     padding-top: 50px;
@@ -185,8 +186,7 @@ export default class Channels extends Mixins(NavigationMixin) {
   &__wrapper {
     height: 100%;
     width: 100%;
-    overflow-y: auto;
-    @extend .themed-scrollbar;
+    overflow-y: visible;
   }
 
   &__header {